requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-debug-toolbar==1.10.1
django-heroku==0.3.1
django-honeypot==0.5.0
django-taggit==0.23.0
easy-thumbnails==2.3
gunicorn==19.9.0
Pillow==3.4.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dotenv==0.9.1
pytz==2018.5
sqlparse==0.2.1
whitenoise==4.1
